# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuultimate
  version: 1.0.0
paths:
...
  /v1/access/check:
    post:
      tags:
      - access
      summary: Check access permission
      operationId: check_access_v1_access_check_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/access/policies:
    post:
      tags:
      - access
      summary: Create access policy
      operationId: create_policy_v1_access_policies_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicySchema'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/access/roles/assign:
    post:
      tags:
      - access
      summary: Assign role to user
      operationId: assign_role_v1_access_roles_assign_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
...
  /v1/backup/snapshots:
    post:
      tags:
      - backup
      summary: Create backup snapshot
      operationId: create_snapshot_v1_backup_snapshots_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/backup/restore:
    post:
      tags:
      - backup
      summary: Restore from snapshot
      operationId: restore_v1_backup_restore_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/backup/integrity-check:
    post:
      tags:
      - backup
      summary: Run integrity check
      operationId: check_integrity_v1_backup_integrity_check_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrityCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrityCheckResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
...
components:
  schemas:
    AccessCheckRequest:
      properties:
        user_id:
          type: string
          title: User Id
          description: User UUID to check access for
        resource:
          type: string
          title: Resource
          description: Resource identifier
          examples:
          - vault:secrets
        action:
          type: string
          title: Action
          description: Action to authorize
          examples:
          - read
      type: object
      required:
      - user_id
      - resource
      - action
      title: AccessCheckRequest
    AccessCheckResponse:
      properties:
        allowed:
          type: boolean
          title: Allowed
          description: Whether access is granted
        reason:
          type: string
          title: Reason
          description: Explanation of the access decision
          default: ''
      type: object
      required:
      - allowed
      title: AccessCheckResponse
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error message
        code:
          type: string
          title: Code
          description: Machine-readable error code
          default: ZUUL_ERROR
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional error context
      type: object
      required:
      - error
      title: ErrorResponse
...
    IntegrityCheckRequest:
      properties:
        target:
          type: string
          title: Target
          description: Target to check integrity of
          examples:
          - identity_db
      type: object
      required:
      - target
      title: IntegrityCheckRequest
    IntegrityCheckResponse:
      properties:
        id:
          type: string
          title: Id
          description: Integrity check UUID
        target:
          type: string
          title: Target
          description: Target that was checked
        passed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Passed
          description: Whether integrity check passed
        status:
          type: string
          title: Status
          description: Check status (completed/failed)
      type: object
      required:
      - id
      - target
      - status
      title: IntegrityCheckResponse
...
    PolicyResponse:
      properties:
        id:
          type: string
          title: Id
          description: Policy UUID
        name:
          type: string
          title: Name
          description: Policy display name
        effect:
          type: string
          title: Effect
          description: Policy effect (allow/deny)
        resource_pattern:
          type: string
          title: Resource Pattern
          description: Resource glob pattern
        action_pattern:
          type: string
          title: Action Pattern
          description: Action glob pattern
        priority:
          type: integer
          title: Priority
          description: Evaluation priority
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
          description: Scoped role UUID
      type: object
      required:
      - id
      - name
      - effect
      - resource_pattern
      - action_pattern
      - priority
      title: PolicyResponse
    PolicySchema:
      properties:
        name:
          type: string
          title: Name
          description: Policy display name
          examples:
          - allow-vault-read
        effect:
          type: string
          title: Effect
          description: Policy effect (allow/deny)
        resource_pattern:
          type: string
          title: Resource Pattern
          description: Resource glob pattern
          examples:
          - vault:*
        action_pattern:
          type: string
          title: Action Pattern
          description: Action glob pattern
          examples:
          - read
        priority:
          type: integer
          title: Priority
          description: Evaluation priority (higher wins)
          default: 0
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
          description: Optional role UUID to scope policy to
      type: object
      required:
      - name
      - effect
      - resource_pattern
      - action_pattern
      title: PolicySchema
...
    RestoreRequest:
      properties:
        snapshot_id:
          type: string
          title: Snapshot Id
          description: Snapshot UUID to restore from
        target:
          type: string
          title: Target
          description: Restore target
          examples:
          - identity_db
      type: object
      required:
      - snapshot_id
      - target
      title: RestoreRequest
    RestoreResponse:
      properties:
        id:
          type: string
          title: Id
          description: Restore job UUID
        snapshot_id:
          type: string
          title: Snapshot Id
          description: Source snapshot UUID
        target:
          type: string
          title: Target
          description: Restore target
        status:
          type: string
          title: Status
          description: Job status (completed/failed)
      type: object
      required:
      - id
      - snapshot_id
      - target
      - status
      title: RestoreResponse
...
    RoleAssignRequest:
      properties:
        role_id:
          type: string
          title: Role Id
          description: Role UUID to assign
        user_id:
          type: string
          title: User Id
          description: User UUID to assign role to
        assigned_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned By
          description: Admin user who made the assignment
      type: object
      required:
      - role_id
      - user_id
      title: RoleAssignRequest
    RoleAssignResponse:
      properties:
        id:
          type: string
          title: Id
          description: Assignment UUID
        role_id:
          type: string
          title: Role Id
          description: Assigned role UUID
        user_id:
          type: string
          title: User Id
          description: User UUID
        assigned_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned By
          description: Admin who made the assignment
      type: object
      required:
      - id
      - role_id
      - user_id
      title: RoleAssignResponse
...
    SnapshotCreate:
      properties:
        name:
          type: string
          title: Name
          description: Snapshot display name
          examples:
          - daily-2026-02-21
        source:
          type: string
          title: Source
          description: Data source to snapshot
          examples:
          - identity_db
      type: object
      required:
      - name
      - source
      title: SnapshotCreate
    SnapshotResponse:
      properties:
        id:
          type: string
          title: Id
          description: Snapshot UUID
        name:
          type: string
          title: Name
          description: Snapshot display name
        source:
          type: string
          title: Source
          description: Data source that was snapshotted
        checksum:
          type: string
          title: Checksum
          description: SHA-256 checksum of snapshot data
        status:
          type: string
          title: Status
          description: Snapshot status (completed/failed)
      type: object
      required:
      - id
      - name
      - source
      - checksum
      - status
      title: SnapshotResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer